1.locationService添加gpsCityId 2.个人中心UI及跳转修改 3.订单提交后清除该商店购物车

+++ app/src/main/java/com/alkaid/trip51/model/shop/Shop.java
@@ -6,7 +6,7 @@ import java.io.Serializable;
  * Created by alkaid on 2015/11/29.
  */
 public class Shop implements Serializable {
-    private int shopid;
+    private long shopid;
     private String shopname;
     private String avgpay;
     private String imageurl;
@@ -18,11 +18,11 @@ public class Shop implements Serializable {
     private int hallstatus;
     private int distance;
 
-    public int getShopid() {
+    public long getShopid() {
         return shopid;
     }
 
-    public void setShopid(int shopid) {
+    public void setShopid(long shopid) {
         this.shopid = shopid;
     }
